Ignore secured remote tests in FactoryTest
These tests need the JVM to use a specific trust store.
For this to work, the sysprop must be set early (when surefire creates the JVM).

But, the trust store is used for all tests, including those which trigger connections to Maven Central.
Therefore, the suite fails on a fresh local repository (because the usual Vert.x dependencies can't be downloaded).